Added rendering Map to Renderer, some small changes in Map

+++ source/Config.hpp
@@ -23,7 +23,7 @@ namespace constants
 	{
 		constexpr auto ROTATION_MODES = 4;
 		constexpr auto STRAIGHT_ANGLE = 90;
-		constexpr pi::coord_t<float> CELL_DIMENSIONS = { 64,64 };
+		constexpr pi::coord_t<int8_t> CELL_DIMENSIONS = { 64,64 };
 	}
 
 	namespace error
